pyproject.toml
[project]
name = "finsysapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.133.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
